requirements.txt
opencv-python
grad-cam
timm
scikit-learn
matplotlib
prettytable
pillow
thop
rfconv